Title: 사칙연산 함수, Time: 43ms, Memory: 16MB, Status: Passed - CodeTree

+++ 250319/사칙연산 함수/quadratic-operations-function.py
@@ -12,7 +12,7 @@ def formula(a, o, c):
         val = a - c
         print(f"{a} {o} {c} = {val}")
     elif o == '/':
-        val = a / c
+        val = int(a / c)
         print(f"{a} {o} {c} = {val}")
     elif o == '*':
         val = a * c
